Adds forcing module (Phase A refactor)
Moves dataset I/O and forcing logic from driver/main.c into a new
internal module src/forcing/:

- src/forcing/rdyforcing.c: lifecycle (Create/Setup/Apply/Destroy),
  CLI option parsing, and dispatch to RDycore source/BC setters
- src/forcing/rdyforcing_dataset.c: binary PETSc Vec I/O for
  constant, homogeneous, raster, and unstructured datasets
- src/forcing/rdyforcing_map.c: nearest-neighbor spatial mapping
  between dataset grids and the RDycore mesh
- include/private/rdyforcingimpl.h: internal types and declarations

RDyForcing is embedded in struct _p_RDy (rdycoreimpl.h) and its
lifecycle is managed automatically by RDySetup, RDyAdvance, and
RDyDestroy. No public API is exposed. driver/main.c is simplified
to a plain create/setup/advance/destroy loop.
