+++ src/forcing/CMakeLists.txt
@@ -1,0 +1,2 @@
+# src/forcing/CMakeLists.txt -- Forcing module sources
+# These are included in the parent CMakeLists.txt via the RDYCORE_C_SOURCE_FILES list.
