+++ src/rdycore.c
@@ -264,6 +264,9 @@ PetscErrorCode RDyDestroy(RDy *rdy) {
     PetscCall(DestroyOperator(&(*rdy)->operator));
   }
 
+  // destroy forcing
+  if ((*rdy)->forcing) PetscCall(RDyDestroyForcing(&(*rdy)->forcing));
+
   // destroy time series
   PetscCall(DestroyTimeSeries(*rdy));
